Rename Button component to NextSceneBtn for better clarity
